+++ circonus/resource_circonus_graph_test.go
@@ -22,7 +22,61 @@ func TestAccCirconusGraph_basic(t *testing.T) {
 		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckDestroyCirconusGraph,
 		Steps: []resource.TestStep{
-			{
+			{ // Create a graph without some settings.
+				Config: fmt.Sprintf(testAccCirconusGraphRemoveConfigFmt,
+					graphCheckName,
+					testAccBroker1,
+					graphName,
+					" foo",
+				),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "name", graphName),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "description", "Terraform Test: mixed graph foo"),
+					resource.TestCheckNoResourceAttr("circonus_graph.mixed-points", "notes"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "graph_style", "line"),
+					resource.TestCheckNoResourceAttr("circonus_graph.mixed-points", "left.%"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "right.%", "2"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "right.logarithmic", "10"),
+					resource.TestCheckNoResourceAttr("circonus_graph.mixed-points", "right.max"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "right.min", "-1"),
+
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "line_style", "stepped"),
+
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.#", "3"),
+
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.caql", ""),
+					resource.TestCheckResourceAttrSet("circonus_graph.mixed-points", "metric.0.check"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.metric_name", "maximum"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.metric_type", "numeric"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.name", "Maximum Latency"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.axis", "left"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.color", "#657aa6"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.function", "gauge"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.active", "true"),
+
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.caql", ""),
+					resource.TestCheckResourceAttrSet("circonus_graph.mixed-points", "metric.1.check"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.metric_name", "minimum"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.metric_type", "numeric"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.name", "Minimum Latency"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.axis", "right"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.color", "#657aa6"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.function", "gauge"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.active", "true"),
+
+					resource.TestCheckResourceAttrSet("circonus_graph.mixed-points", "metric.2.caql"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.check", ""),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.metric_type", "caql"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.name", "Foo sum CAQL"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.axis", "left"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.color", "#657aa6"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.function", "gauge"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.active", "true"),
+
+					resource.TestCheckNoResourceAttr("circonus_graph.mixed-points", "tags.#"),
+				),
+			},
+			{ // Add those settings.
 				Config: fmt.Sprintf(testAccCirconusGraphConfigFmt,
 					graphCheckName,
 					testAccBroker1,
@@ -134,6 +188,60 @@ func TestAccCirconusGraph_basic(t *testing.T) {
 					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "tags.#", "2"),
 					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "tags.0", "author:terraform"),
 					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "tags.1", "lifecycle:unittest"),
+				),
+			},
+			{ // force removal of settings to verify they are removed from the graph.
+				Config: fmt.Sprintf(testAccCirconusGraphRemoveConfigFmt,
+					graphCheckName,
+					testAccBroker1,
+					graphName,
+					" foo",
+				),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "name", graphName),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "description", "Terraform Test: mixed graph foo"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "notes", ""),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "graph_style", "line"),
+					resource.TestCheckNoResourceAttr("circonus_graph.mixed-points", "left.%"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "right.%", "2"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "right.logarithmic", "10"),
+					resource.TestCheckNoResourceAttr("circonus_graph.mixed-points", "right.max"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "right.min", "-1"),
+
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "line_style", "stepped"),
+
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.#", "3"),
+
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.caql", ""),
+					resource.TestCheckResourceAttrSet("circonus_graph.mixed-points", "metric.0.check"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.metric_name", "maximum"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.metric_type", "numeric"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.name", "Maximum Latency"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.axis", "left"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.color", "#657aa6"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.function", "gauge"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.0.active", "true"),
+
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.caql", ""),
+					resource.TestCheckResourceAttrSet("circonus_graph.mixed-points", "metric.1.check"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.metric_name", "minimum"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.metric_type", "numeric"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.name", "Minimum Latency"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.axis", "right"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.color", "#657aa6"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.function", "gauge"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.1.active", "true"),
+
+					resource.TestCheckResourceAttrSet("circonus_graph.mixed-points", "metric.2.caql"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.check", ""),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.metric_type", "caql"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.name", "Foo sum CAQL"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.axis", "left"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.color", "#657aa6"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.function", "gauge"),
+					resource.TestCheckResourceAttr("circonus_graph.mixed-points", "metric.2.active", "true"),
+
+					resource.TestCheckNoResourceAttr("circonus_graph.mixed-points", "tags.#"),
 				),
 			},
 		},
@@ -268,5 +376,86 @@ EOF
   }
 
   tags = "${var.test_tags}"
+}
+`
+
+const testAccCirconusGraphRemoveConfigFmt = `
+variable "test_tags" {
+  type = list(string)
+  default = [ "author:terraform", "lifecycle:unittest" ]
+}
+
+resource "circonus_check" "api_latency" {
+  active = true
+  name = "%s"
+  period = "60s"
+
+  collector {
+    id = "%s"
+  }
+
+  icmp_ping {
+    count = 5
+  }
+
+  metric {
+    name = "maximum"
+    type = "numeric"
+  }
+
+  metric {
+    name = "minimum"
+    type = "numeric"
+  }
+
+  tags = "${var.test_tags}"
+  target = "api.circonus.com"
+}
+
+resource "circonus_graph" "mixed-points" {
+  name = "%s"
+  description = "Terraform Test: mixed graph%s"
+
+  right = {
+    logarithmic = 10
+    min = -1
+  }
+
+  metric {
+    check = "${circonus_check.api_latency.checks[0]}"
+    metric_name = "maximum"
+    metric_type = "numeric"
+    name = "Maximum Latency"
+    axis = "left" # right
+    color = "#657aa6"
+    function = "gauge"
+    active = true
+  }
+
+  metric {
+    check = "${circonus_check.api_latency.checks[0]}"
+    metric_name = "minimum"
+    metric_type = "numeric"
+    name = "Minimum Latency"
+    axis = "right" # left
+    color = "#657aa6"
+    function = "gauge"
+    active = true
+  }
+
+  metric {
+    active = true
+    axis = "left"
+    caql = <<-EOF
+      find:average("foo")
+      | stats:sum()
+      | label("foo_sum")
+EOF
+    color = "#657aa6"
+    function = "gauge"
+    legend_formula = "=round(VAL,2)"
+    metric_type = "caql"
+    name = "Foo sum CAQL"
+  }
 }
 `